Rename SwellSampleMapping -> DrumSampleMapping
