+++ osu.Game.Rulesets.Taiko/Audio/DrumSampleMapping.cs
@@ -5,17 +5,29 @@ using System;
 using osu.Framework.Audio;
 using osu.Framework.Audio.Sample;
 using osu.Game.Audio;
+using osu.Game.Beatmaps.ControlPoints;
 
-namespace osu.Game.Rulesets.Taiko.Objects
+namespace osu.Game.Rulesets.Taiko.Audio
 {
-    public class SwellSampleMapping : IComparable<SwellSampleMapping>
+    public class DrumSampleMapping : IComparable<DrumSampleMapping>
     {
         public double Time;
-        public SampleInfo Centre;
-        public SampleInfo Rim;
+        public readonly SampleInfo Centre;
+        public readonly SampleInfo Rim;
 
         public SampleChannel CentreChannel { get; private set; }
         public SampleChannel RimChannel { get; private set; }
+
+        public DrumSampleMapping()
+        {
+        }
+
+        public DrumSampleMapping(SampleControlPoint samplePoint)
+        {
+            Time = samplePoint.Time;
+            Centre = samplePoint.GetSampleInfo();
+            Rim = samplePoint.GetSampleInfo(SampleInfo.HIT_CLAP);
+        }
 
         public void RetrieveChannels(AudioManager audio)
         {
@@ -23,6 +35,6 @@ namespace osu.Game.Rulesets.Taiko.Objects
             RimChannel = Rim.GetChannel(audio.Sample);
         }
 
-        public int CompareTo(SwellSampleMapping other) => Time.CompareTo(other.Time);
+        public int CompareTo(DrumSampleMapping other) => Time.CompareTo(other.Time);
     }
 }
