+++ osu.Game.Rulesets.Taiko/Objects/Swell.cs
@@ -7,6 +7,7 @@ using osu.Game.Audio;
 using osu.Game.Beatmaps;
 using osu.Game.Beatmaps.ControlPoints;
 using osu.Game.Rulesets.Objects.Types;
+using osu.Game.Rulesets.Taiko.Audio;
 
 namespace osu.Game.Rulesets.Taiko.Objects
 {
@@ -21,25 +22,17 @@ namespace osu.Game.Rulesets.Taiko.Objects
         /// </summary>
         public int RequiredHits = 10;
 
-        public List<SwellSampleMapping> ProgressionSamples = new List<SwellSampleMapping>();
+        public List<DrumSampleMapping> ProgressionSamples = new List<DrumSampleMapping>();
 
         protected override void ApplyDefaultsToSelf(ControlPointInfo controlPointInfo, BeatmapDifficulty difficulty)
         {
             base.ApplyDefaultsToSelf(controlPointInfo, difficulty);
 
-            var progressionSamplePoints =
-                new[] { controlPointInfo.SamplePointAt(StartTime) }
+            var progressionSamplePoints = new[] { controlPointInfo.SamplePointAt(StartTime) }
                 .Concat(controlPointInfo.SamplePoints.Where(p => p.Time > StartTime && p.Time <= EndTime));
 
             foreach (var point in progressionSamplePoints)
-            {
-                ProgressionSamples.Add(new SwellSampleMapping
-                {
-                    Time = point.Time,
-                    Centre = point.GetSampleInfo(),
-                    Rim = point.GetSampleInfo(SampleInfo.HIT_CLAP)
-                });
-            }
+                ProgressionSamples.Add(new DrumSampleMapping(point));
         }
     }
 }
